Refactored endpoints for the YearlyStatistics

+++ GentlemensClub/Models/YearlyStatistics/Months.cs
@@ -7,5 +7,5 @@ public class Months
     public float High { get; set; }
     public float Low { get; set; }
     public float Close { get; set; }
-    public float Volume { get; set; }
+    public int Volume { get; set; }
 }
